+++ src/main/java/com/github/manasmods/manascore/api/data/gen/ItemTagProvider.java
@@ -5,15 +5,32 @@
 package com.github.manasmods.manascore.api.data.gen;
 
 import net.minecraft.data.tags.ItemTagsProvider;
+import net.minecraft.tags.TagKey;
+import net.minecraft.world.item.Item;
 import net.minecraftforge.data.event.GatherDataEvent;
 import org.jetbrains.annotations.ApiStatus.AvailableSince;
+import org.jetbrains.annotations.ApiStatus.NonExtendable;
 import org.jetbrains.annotations.ApiStatus.OverrideOnly;
+
+import java.util.Arrays;
+import java.util.function.Supplier;
 
 @AvailableSince("1.0.0.0")
 public abstract class ItemTagProvider extends ItemTagsProvider {
 
     public ItemTagProvider(GatherDataEvent gatherDataEvent, String modId, BlockTagProvider blockTagProvider) {
         super(gatherDataEvent.getGenerator(), blockTagProvider, modId, gatherDataEvent.getExistingFileHelper());
+    }
+
+    @SafeVarargs
+    @NonExtendable
+    protected final void subTag(TagKey<Item> root, TagKey<Item> sub, Supplier<? extends Item>... items) {
+        subTag(root, sub, Arrays.stream(items).map(Supplier::get).toArray(Item[]::new));
+    }
+
+    private void subTag(TagKey<Item> root, TagKey<Item> sub, Item... items) {
+        tag(root).addTag(sub);
+        tag(sub).add(items);
     }
 
     @Override